// src/app/shared/interfaces/film.interface.ts
export class IFilm {
    imdbID: string;
    imdbRating: string;
    title: string;
    actors: string;
    awards: string;
    boxOffice: string;
    country: string;
    director: string;
    genre: string;
    language: string;
    metascore: string;
    plot: string;
    poster: string;
    production: string;
    rated: string;
    released: string;
    runtime: string;
}

export class IFilmResponse {
    imdbID: string;
    imdbRating: string;
    Title: string;
    Actors: string;
    Awards: string;
    BoxOffice: string;
    Country: string;
    Director: string;
    Genre: string;
    Language: string;
    Metascore: string;
    Plot: string;
    Poster: string;
    Production: string;
    Rated: string;
    Released: string;
    Runtime: string;
}